Split module tests/helpers/application.js into start-app.js et destroy-app.js to make bumping easier (in new Ember app, they are splitted)
