+++ live/tests/acceptance/b7-epreuve-points-communs-test.js
@@ -5,7 +5,8 @@ import {
   afterEach
 } from 'mocha';
 import { expect } from 'chai';
-import { startApp, destroyApp } from '../helpers/application';
+import startApp from '../helpers/start-app';
+import destroyApp from '../helpers/destroy-app';
 import $ from 'jquery';
 
 describe('Acceptance | b7 - Points communs a toutes les épreuves | ', function() {
