+++ live/tests/acceptance/compte-authentication-and-profile-test.js
@@ -5,7 +5,8 @@ import {
   it
 } from 'mocha';
 import { expect } from 'chai';
-import { startApp, destroyApp } from '../helpers/application';
+import startApp from '../helpers/start-app';
+import destroyApp from '../helpers/destroy-app';
 import {
   authenticateAsSimpleUser,
   authenticateAsPrescriber
