+++ live/tests/acceptance/index-test.js
@@ -5,7 +5,8 @@ import {
   it
 } from 'mocha';
 import { expect } from 'chai';
-import { destroyApp, startApp } from '../helpers/application';
+import startApp from '../helpers/start-app';
+import destroyApp from '../helpers/destroy-app';
 
 describe('Acceptance | index', function() {
   let application;
