+++ live/tests/acceptance/j2-compare-answer-solution-qroc-test.js
@@ -5,7 +5,8 @@ import {
   afterEach
 } from 'mocha';
 import { expect } from 'chai';
-import { startApp, destroyApp } from '../helpers/application';
+import startApp from '../helpers/start-app';
+import destroyApp from '../helpers/destroy-app';
 import $ from 'jquery';
 
 describe('Acceptance | j2 - Comparer réponses et solutions pour un QROC | ', function() {
