+++ live/tests/acceptance/page-accueil-test.js
@@ -5,7 +5,8 @@ import {
   it
 } from 'mocha';
 import { expect } from 'chai';
-import { startApp, destroyApp } from '../helpers/application';
+import startApp from '../helpers/start-app';
+import destroyApp from '../helpers/destroy-app';
 
 describe('Acceptance | La page d\'accueil', function() {
