+++ live/tests/helpers/destroy-app.js
@@ -1,22 +1,6 @@
-import Application from '../../app';
-import config from '../../config/environment';
 import { run } from '@ember/runloop';
-import { merge } from '@ember/polyfills';
-import './responsive';
 
-export function startApp(attrs) {
-  let attributes = merge({}, config.APP);
-  attributes = merge(attributes, attrs); // use defaults, but you can override;
-
-  return run(() => {
-    const application = Application.create(attributes);
-    application.setupForTesting();
-    application.injectTestHelpers();
-    return application;
-  });
-}
-
-export function destroyApp(application) {
+export default function destroyApp(application) {
   run(application, 'destroy');
   if (window.server) {
     window.server.shutdown();
